requirements.txt
pycups==2.0.4
flask==2.0.3
flask-sqlalchemy==2.5.1
requests==2.21.0
pillow==5.4.1
